add and fix categories feature

+++ CodeReviews.Console.Phonebook/STUDY.ConsoleApp.Phonebook/Services/CategoryService.cs
@@ -4,7 +4,7 @@ using STUDY.ConsoleApp.Phonebook.Models;
 
 namespace STUDY.ConsoleApp.Phonebook.Services;
 
-public class CategoryService
+public static class CategoryService
 {
     public static void AddCategory()
     {
@@ -12,8 +12,12 @@ public class CategoryService
         {
             Name = AnsiConsole.Ask<string>("Enter the category name: ")
         };
-        AnsiConsole.MarkupLine("[green]Category added successfully![/]");
-        CategoryController.AddCategory(category);
+
+        AnsiConsole.MarkupLine(CategoryController.AddCategory(category)
+            ? "[green]Category added successfully![/]"
+            : "[red]Category already exists![/]");
+        Console.WriteLine("Press any key to continue...");
+        Console.ReadLine();
     }
 
     public static void ListAllCategories()
@@ -21,10 +25,18 @@ public class CategoryService
         var categories = CategoryController.ListAllCategories();
         UserInterface.ShowCategoriesTable(categories);
     }
-    
-    public static Category GetCategoryOptionInput()
+
+    private static Category? GetCategoryOptionInput()
     {
         var categories = CategoryController.ListAllCategories();
+        if (categories.Count == 0)
+        {
+            AnsiConsole.MarkupLine("[red]No categories found.[/]");
+            Console.WriteLine("Press any key to continue...");
+            Console.ReadLine();
+            return null;
+        }
+
         var categoriesNames = categories.Select(x => x.Name);
 
         var selectedOption = AnsiConsole.Prompt(new SelectionPrompt<string>()
@@ -35,15 +47,30 @@ public class CategoryService
         return category;
     }
 
+    public static List<Category>? GetMultipleCategoriesInput()
+    {
+        var categories = CategoryController.ListAllCategories();
+        if (categories.Count == 0)
+            return null;
+
+        var categoriesNames = categories.Select(x => x.Name);
+        var selectedOptions = AnsiConsole.Prompt(new MultiSelectionPrompt<string>()
+            .Title("Select one or more categories:")
+            .AddChoices(categoriesNames)
+        );
+        var categoriesList = categories.Where(x => selectedOptions.Contains(x.Name)).ToList();
+        return categoriesList;
+    }
+
     public static void UpdateCategory()
     {
         var category = GetCategoryOptionInput();
-        if (AnsiConsole.Confirm("Do you want to update the category's name?"))
-        {
-            category.Name = AnsiConsole.Ask<string>("Enter the new category's name: ");
-            CategoryController.UpdateCategory(category);
-            AnsiConsole.MarkupLine("[green]Category updated successfully![/]");
-        }
+        if (category == null) return;
+
+        category.Name = AnsiConsole.Ask<string>("Enter the new category's name: ");
+        CategoryController.UpdateCategory(category);
+        AnsiConsole.MarkupLine("[green]Category updated successfully![/]");
+
         Console.WriteLine("Press any key to continue...");
         Console.ReadLine();
     }
@@ -51,6 +78,9 @@ public class CategoryService
     public static void DeleteCategory()
     {
         var category = GetCategoryOptionInput();
+        if (category == null) return;
+
+        if (!AnsiConsole.Confirm("Are you sure you want to delete this category?")) return;
         CategoryController.DeleteCategory(category);
         AnsiConsole.MarkupLine("[green]Category deleted successfully![/]");
         Console.WriteLine("Press any key to continue...");
